`enu2rgaz` and `get_enu_vector_ecef` moved to `geometry_utils.py`
